fix: Restructure landing page folders and adjust related components
