+++ frontend/src/app/(internal)/landing-page.tsx
